More bugs from the Giant King miniboss encounter:
Because the turn flow made a detour during the turn, disripting natural turn flow progression, if the boss was killed in one shot, it was entering the "encounter phase" even though the encounter was technically over. To fix this, I had to make the actual combat take place during the next turn. Otherwise, if the boss dies in one hit during his setup, he died in a "non encounter phase" which makes things wonky.