Added the ability to view an Ascii map of the party's positions on the grid. It also displays each member's HP and if they're currently engaged in combat.
Made it so that XP isnt split during a shared encounter- let's see how that goes... lol

Fixed a bug preventing mobs from despawning after a Warrior runs away after using Taunt and populating the encounter playerArray with duplicates.

Cached the amountOfCharacters at the start of the game, to prevent the gameEngine playerArray from becoming populated by hostile mobs. The issue this caused, is that the combat encounters scale based on how many players are in the game, and it was accounting for mobs on the field as well. This allowed like 20 mobs to spawn if there were 4 players each doing their own combats.

Decreased the effect of luck when checking for item drops.

Reduced chain lightning damage